requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
brotli==1.2.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.13.1
croniter==6.0.0
diff-match-patch==20241021
Django==6.0.1
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-import-export==4.3.14
django-redis==6.0.0
django-rq==3.2.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.2
fsspec==2025.12.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
mutagen==1.47.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
packaging==25.0
pillow==12.1.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
regex==2025.11.3
requests==2.32.5
rq==2.6.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
tablib==3.9.0
tenacity==9.1.2
tiktoken==0.12.0
# torch==2.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==15.0.1
whitenoise==6.11.0
